backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
rapidfuzz
sendgrid
web3
supabase
httpx